requirements.txt
redis
python-socketio
aiohttp
python-dotenv